* moved dotprod test case from short to float * update compileC.sh so that it can be used to compile generated files * add a simple sse.h macro header
git-svn-id: http://svn.cri.ensmp.fr/svn/validation/trunk@376 67b5017d-e62d-0410-a871-f72b1c62810f

+++ SAC/kernels/DOTPROD/DOTPROD.c
@@ -1,8 +1,8 @@
 #include<math.h>
-short dotprod(short b[1000000], short c[1000000])
+float dotprod(float b[1000000], float c[1000000])
 {
     int i;
-    short a=0;
+    float a=0;
     for(i=0;i<1000000;++i)
     {
         a = a + b[i] * c[i] ;
